+++ spec/plans/init_spec.rb
@@ -28,14 +28,14 @@ describe 'facts' do
       expect_task('facts').always_return(fact_output)
       inventory.expects(:add_facts).with(target, fact_output).returns(fact_output)
 
-      expect(run_plan('facts', 'nodes' => [node]).value).to eq(results(fact_output))
+      expect(run_plan('facts', 'targets' => [node]).value).to eq(results(fact_output))
     end
 
     it 'does not mask errors' do
       expect_task('facts').always_return(err_output)
       inventory.expects(:add_facts).never
 
-      expect(run_plan('facts', 'nodes' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
+      expect(run_plan('facts', 'targets' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
     end
   end
 
@@ -46,14 +46,14 @@ describe 'facts' do
       expect_task('facts').always_return(fact_output)
       inventory.expects(:add_facts).with(target, fact_output).returns(fact_output)
 
-      expect(run_plan('facts', 'nodes' => [node]).value).to eq(results(fact_output))
+      expect(run_plan('facts', 'targets' => [node]).value).to eq(results(fact_output))
     end
 
     it 'does not mask errors' do
       expect_task('facts').always_return(err_output)
       inventory.expects(:add_facts).never
 
-      expect(run_plan('facts', 'nodes' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
+      expect(run_plan('facts', 'targets' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
     end
   end
 
@@ -64,14 +64,14 @@ describe 'facts' do
       expect_task('facts').always_return(fact_output)
       inventory.expects(:add_facts).with(target, fact_output).returns(fact_output)
 
-      expect(run_plan('facts', 'nodes' => [node]).value).to eq(results(fact_output))
+      expect(run_plan('facts', 'targets' => [node]).value).to eq(results(fact_output))
     end
 
     it 'does not mask errors' do
       expect_task('facts').always_return(err_output)
       inventory.expects(:add_facts).never
 
-      expect(run_plan('facts', 'nodes' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
+      expect(run_plan('facts', 'targets' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
     end
   end
 
@@ -82,14 +82,14 @@ describe 'facts' do
       expect_task('facts').always_return(fact_output)
       inventory.expects(:add_facts).with(target, fact_output).returns(fact_output)
 
-      expect(run_plan('facts', 'nodes' => [node]).value).to eq(results(fact_output))
+      expect(run_plan('facts', 'targets' => [node]).value).to eq(results(fact_output))
     end
 
     it 'does not mask errors' do
       expect_task('facts').always_return(err_output)
       inventory.expects(:add_facts).never
 
-      expect(run_plan('facts', 'nodes' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
+      expect(run_plan('facts', 'targets' => [node]).value.msg).to eq("Plan aborted: run_task 'facts' failed on 1 target")
     end
   end
 
@@ -106,7 +106,7 @@ describe 'facts' do
       result_set = Bolt::ResultSet.new(
         nodes.map { |node| Bolt::Result.new(Bolt::Target.new(node), value: fact_output(node)) }
       )
-      expect(run_plan('facts', 'nodes' => nodes).value).to eq(result_set)
+      expect(run_plan('facts', 'targets' => nodes).value).to eq(result_set)
     end
 
     it 'does not mask errors' do
@@ -114,7 +114,7 @@ describe 'facts' do
       expect_task('facts').return_for_targets(target_results)
       inventory.expects(:add_facts).never
 
-      expect(run_plan('facts', 'nodes' => nodes).value.msg).to eq("Plan aborted: run_task 'facts' failed on 3 targets")
+      expect(run_plan('facts', 'targets' => nodes).value.msg).to eq("Plan aborted: run_task 'facts' failed on 3 targets")
     end
   end
 end